Detect network address on Node 18
In the output of `os.networkInterfaces()`, the `family` field in Node 18 has been changed from a string (`"IPv4"`) to an int (`4`). It’s an undocumented change ([there is an issue about it](https://github.com/nodejs/node/issues/42787)) but it’s there in Node v18.1.0. 

This change supports both version of this field, improving compatibility.